requirements.txt
Flask
langchain 
openai
langchain-openai
python-dotenv

